Update to Monogame 3.6, using Monogame Pipeline Tool for content

+++ MonoGameBallerburg/MonoGameBallerburg/Manager/ScreenManager.cs
@@ -405,7 +405,7 @@ namespace MonoGameBallerburg
       spriteBatch.Draw(
           blankTexture,
           new Rectangle(0, 0, viewport.Width, viewport.Height),
-          new Color(0, 0, 0, (byte)alpha));
+          new Color((byte)0, (byte)0, (byte)0, (byte)alpha));
 
       spriteBatch.End();
     }
