+++ MonoGameBallerburg/MonoGameBallerburg/Screens/MessageBoxScreen.cs
@@ -141,7 +141,7 @@ namespace MonoGameBallerburg.Screens
       var backgroundRectangle = new Rectangle(160, 90, 320, 300);
 
       // Fade the popup alpha during transitions.
-      var color = new Color(255, 255, 255, TransitionAlpha);
+      var color = new Color((byte)255, (byte)255, (byte)255, TransitionAlpha);
 
       spriteBatch.Begin();
